+++ plugin-maven/src/main/java/com/diffplug/spotless/maven/java/Java.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.diffplug.maven.spotless;
+package com.diffplug.spotless.maven.java;
 
 import static java.util.stream.Collectors.toList;
 
@@ -28,6 +28,9 @@ import com.diffplug.spotless.FormatExceptionPolicyStrict;
 import com.diffplug.spotless.Formatter;
 import com.diffplug.spotless.FormatterStep;
 import com.diffplug.spotless.LineEnding;
+import com.diffplug.spotless.maven.FormatterFactory;
+import com.diffplug.spotless.maven.FormatterStepFactory;
+import com.diffplug.spotless.maven.MojoConfig;
 
 public class Java implements FormatterFactory {
