+++ plugin-maven/src/test/java/com/diffplug/spotless/maven/java/EclipseFormatStepTest.java
@@ -13,36 +13,30 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.diffplug.maven.spotless;
+package com.diffplug.spotless.maven.java;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.Test;
 
-public class ImportOrderTest extends MavenIntegrationTest {
-	@Test
-	public void file() throws Exception {
-		write("import.properties", getTestResource("java/importsorter/import.properties"));
-		writePomWithJavaSteps(
-				"<importOrder>",
-				"  <file>${basedir}/import.properties</file>",
-				"</importOrder>");
-		runTest();
-	}
+import com.diffplug.spotless.maven.MavenIntegrationTest;
+
+public class EclipseFormatStepTest extends MavenIntegrationTest {
 
 	@Test
-	public void order() throws Exception {
+	public void testEclipse() throws Exception {
 		writePomWithJavaSteps(
-				"<importOrder>",
-				"  <order>java,javax,org,\\#com</order>",
-				"</importOrder>");
-		runTest();
-	}
+				"<eclipse>",
+				"  <file>${basedir}/formatter.xml</file>",
+				"  <version>4.7.1</version>",
+				"</eclipse>");
 
-	private void runTest() throws Exception {
-		write("src/main/java/test.java", getTestResource("java/importsorter/JavaCodeUnsortedImports.test"));
+		write("src/main/java/test.java", getTestResource("java/eclipse/format/JavaCodeUnformatted.test"));
+		write("formatter.xml", getTestResource("java/eclipse/format/formatter.xml"));
+
 		mavenRunner().withArguments("spotless:apply").runNoError();
+
 		String actual = read("src/main/java/test.java");
-		assertThat(actual).isEqualTo(getTestResource("java/importsorter/JavaCodeSortedImports.test"));
+		assertThat(actual).isEqualTo(getTestResource("java/eclipse/format/JavaCodeFormatted.test"));
 	}
 }
